add port to .env and update server messages

+++ backend/controllers/productController.js
@@ -37,7 +37,7 @@ export const updateProduct = async (req, res) => {
   const product = req.body;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(404).json({ success: false, message: 'Invalid product ID, product not found' });
+    return res.status(404).json({ success: false, message: 'Invalid product ID' });
   }
 
   try {
@@ -52,11 +52,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, message: 'Invalid product ID' });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: 'Product deleted' });
   } catch (error) {
     console.log('Error in deleting product:', error.message);
-    res.status(404).json({ success: false, message: 'Product not found' });
+    res.status(500).json({ success: false, message: 'Server Error' });
   }
 };
